Fighting functionality added
* Unified bussiness functionality to be coherent
* Removed some deadend functionality and entity attributes
* Rewritten some code after our former members

+++ dndtroops-api/src/main/java/cz/muni/fi/pa165/dndtroops/facade/RoleFacade.java
@@ -64,5 +64,5 @@ public interface RoleFacade {
      * @param role a Role for which to compute random attacking force
      * @return attack value
      */
-    float computeAttackingForce(RoleDTO role);
+    int computeAttackingForce(RoleDTO role);
 }
